Past weather aggregation + junit test

+++ src/main/java/com/escion/climbingweathercollector/dto/openweatherapi/CommonWeather.java
@@ -20,8 +20,8 @@ public abstract class CommonWeather {
     public Double windDeg;
     @JsonProperty("rain")
     @JsonIgnore
-    public Integer rain;
+    public Double rain;
     @JsonProperty("snow")
     @JsonIgnore
-    public Integer snow;
+    public Double snow;
 }
